% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_code.R, R/링geocode 크롤.R
\name{geo_code}
\alias{geo_code}
\alias{geo_code}
\title{geo_code}
\usage{
geo_code(address, b = NULL)

geo_code(address, b = NULL)
}
\arguments{
\item{address}{is character value.}

\item{address}{is character value.}
}
\description{
geo_code

geocode
}
\examples{
geo_code(address=c('hoston texas','hoston texas'))
}
